imported changes from approved PR
