+++ src/components/Footer/footer.scss
@@ -1,0 +1,30 @@
+@import '../../styles/variables/variables.scss';
+
+.footer {
+  font-family: "Poppins-Light", "sans-serif";
+  font-size: 12px;
+  line-height: 16px;
+  font-weight: 300;
+  background-color: $main-color;
+  color: $white;
+  font-size: 14px;
+  line-height: 22px;
+  &__content {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+  }
+}
+
+.menu__footer {
+  &-list {
+    display: flex;
+    list-style-type: none;
+    align-items: baseline;
+  }
+  &-item {
+    display: flex;
+    margin-right: 20px;
+  }
+}
