--- src/componentsBasic/home/home.js
@@ -1,13 +1,0 @@
-import Info from "./homeSections/homeInfo/HomeInfo";
-import Subscribe from "./homeSections/homeSubscribe/HomeSubscribe";
-
-const Home = () => {
-  return (
-    <>
-      <Info />
-      <Subscribe />
-    </>
-  );
-};
-
-export default Home;
