--- src/componentsUniversal/button/button.scss
@@ -1,12 +1,0 @@
-@import '../../stylesUniversal/mixins/mixins.scss';
-@import '../../stylesUniversal/variables/variables.scss';
-
-.btn {
-    @include BTN ($white, 600, $main-color, none, 10px, flex, row, center, center);
-    font-family: "Baloo2-Bold", "sans-serif";
-    font-size: 14px;
-    line-height: 22px;  
-    &:hover {
-      background-color: $button-hover;
-    }
-  }
