--- src/componentsUniversal/dropdown/Dropdown.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-
-const Dropdown = () => {
-    return (
-
-    )
-}
-
-export default Dropdown;
