--- src/componentsUniversal/header/header.scss
@@ -1,34 +1,0 @@
-@import '../../stylesUniversal/mixins/mixins.scss';
-
-.header {
-  font-family: 'Baloo2-Regular', sans-serif;
-  font-size: 14px;
-  line-height: 22px;
-  padding-top: 62px;
-  margin-bottom: 47px;
-  &__top {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-  }
-  &__button {
-    width: 152px;
-    height: 40px;
-  }
-}
-
-.menu {
-  &__link:hover {
-    color: $main-color;
-    font-family: 'Baloo2-SemiBold', sans-serif;
-  }
-  &__list {
-    display: flex;
-    list-style-type: none;
-    align-items: baseline;
-  }
-  &__item {
-    display: flex;
-    margin-left: 25px;
-  }
-}
