+++ src/pages/Home/HomeSections/HomeInfo/homeInfo.scss
