+++ src/pages/Home/HomeSections/HomeInfo/index.js
@@ -1,0 +1,1 @@
+export {default} from './HomeInfo';
