+++ src/pages/Home/HomeSections/HomeSubscribe/homeSubscribe.scss
@@ -1,0 +1,48 @@
+@import '../../../../styles/variables/variables.scss';
+
+.subscribe {
+  font-family: "Poppins-Regular", "sans-serif";
+  font-size: 14px;
+  line-height: 20px;
+  margin-bottom: 51px;
+  &__inner {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+  &__content {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: flex-start;
+    width: 50%;
+    height: 120px;
+  }
+  &__box {
+    &-title {
+      font-family: "Baloo2-Regular", sans-serif;
+      font-size: 28px;
+      line-height: 36px;
+      margin-bottom: 12px;
+    }
+    &-text {
+      margin-bottom: 12px;
+    }
+  }
+  &__img {
+    width: 376px;
+    height: 240px;
+    display: inline-block;
+  }
+  &__button {
+    background: transparent;
+    color: $button-hover;
+    border: 1px $button-hover solid;
+    width: 152px;
+    height: 40px;
+    &:hover {
+      background: $button-hover;
+      color: $white;
+    }
+  }
+}
