+++ src/pages/Home/HomeSections/HomeSubscribe/index.js
@@ -1,0 +1,1 @@
+export {default} from './HomeSubscribe';
